Renamed RssCrawler to PodcastCrawler; Inject list of producers
