+++ inside-java/src/main/kotlin/by/jprof/telegram/opinions/insidejava/Handler.kt
@@ -1,4 +1,4 @@
-package by.jprof.telegram.opinions.insidejava.podcast
+package by.jprof.telegram.opinions.insidejava
 
 import by.jprof.telegram.components.config.componentsDynamoModule
 import by.jprof.telegram.components.config.componentsEnvModule
@@ -8,13 +8,13 @@ import by.jprof.telegram.opinions.insidejava.config.envModule
 import by.jprof.telegram.opinions.insidejava.config.rssModule
 import by.jprof.telegram.opinions.news.config.newsQueueDynamoModule
 import by.jprof.telegram.opinions.news.config.newsQueueEnvModule
+import by.jprof.telegram.opinions.news.produce.Producer
 import com.amazonaws.services.lambda.runtime.Context
 import com.amazonaws.services.lambda.runtime.RequestHandler
 import com.amazonaws.services.lambda.runtime.events.ScheduledEvent
 import kotlinx.coroutines.runBlocking
 import org.koin.core.KoinComponent
 import org.koin.core.context.startKoin
-import org.koin.core.inject
 
 class Handler : RequestHandler<ScheduledEvent, Unit>, KoinComponent {
     init {
@@ -32,10 +32,10 @@ class Handler : RequestHandler<ScheduledEvent, Unit>, KoinComponent {
         }
     }
 
-    private val producer by inject<RssCrawler>()
+    private val producers: List<Producer> = getKoin().getAll()
 
     override fun handleRequest(event: ScheduledEvent, context: Context?) =
-        runBlocking { producer.produce() }
+        runBlocking { producers.forEach { it.produce() } }
 }
 
 fun main() {
